renamed plugin and split into folder structure
