+++ setup.py
@@ -68,7 +68,7 @@ setup(
             "WellCrossSectionFMU = webviz_subsurface.plugins:WellCrossSectionFMU",
             "AssistedHistoryMatchingAnalysis = webviz_subsurface.plugins:AssistedHistoryMatchingAnalysis",
             "WellCompletions = webviz_subsurface.plugins:WellCompletions",
-            "NetworkControlModes = webviz_subsurface.plugins:NetworkControlModes"
+            "NetworkAnalysis = webviz_subsurface.plugins:NetworkAnalysis"
         ],
         "ert": ["webviz_subsurface_jobs = webviz_subsurface.ert_jobs.jobs"],
         "console_scripts": [
