+++ webviz_subsurface/plugins/__init__.py
@@ -57,4 +57,4 @@ from ._well_cross_section import WellCrossSection
 from ._well_cross_section_fmu import WellCrossSectionFMU
 from ._assisted_history_matching_analysis import AssistedHistoryMatchingAnalysis
 from ._well_completions import WellCompletions
-from ._network_control_modes import NetworkControlModes
+from ._network_analysis import NetworkAnalysis
