+++ webviz_subsurface/plugins/_network_analysis/__init__.py
@@ -1,0 +1,1 @@
+from .network_analysis import NetworkAnalysis
