+++ webviz_subsurface/plugins/_network_analysis/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .selections_controllers import selections_controllers
