+++ webviz_subsurface/plugins/_network_analysis/figures/__init__.py
@@ -1,0 +1,1 @@
+from .figures import make_area_graph, make_node_pressure_graph
