+++ webviz_subsurface/plugins/_network_analysis/views/__init__.py
@@ -1,0 +1,1 @@
+from .main_view import main_view
